Moved queryparser to its own module
