+++ src/github.com/cjlucas/unnamedcast/server/queryparser/queryparser.go
@@ -1,4 +1,4 @@
-package main
+package queryparser
 
 import (
 	"fmt"
