+++ src/github.com/cjlucas/unnamedcast/server/queryparser/queryparser_test.go
@@ -1,4 +1,4 @@
-package main
+package queryparser
 
 import (
 	"net/url"
